# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freehold API",
    "version": "0.1.0",
    "description": "Governed memory backend for AI agents, built on the Allod format."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8710",
      "description": "Local Freehold"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token"
      }
    },
    "schemas": {
...
      "GraphInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": ["memory", "repo"]
          },
          "autoPushNotes": {
            "type": "boolean"
          },
          "embedder": {
            "type": "string",
            "enum": ["hash", "semantic"]
          },
          "allodGraphId": {
            "type": "string"
          },
          "originRemote": {
            "type": ["string", "null"]
          }
        },
        "required": [
          "id",
          "name",
          "path",
          "kind",
          "autoPushNotes",
          "embedder",
          "allodGraphId",
          "originRemote"
        ]
      },
      "RegisterGraphBody": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "Absolute path to the repo checkout"
          },
          "id": {
            "type": "string",
            "description": "Registry slug id; defaults to basename of path"
          },
          "name": {
            "type": "string",
            "description": "Display name; defaults to id"
          }
        },
        "required": ["path"]
      },
      "UpdateGraphBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "autoPushNotes": {
            "type": "boolean"
          },
          "embedder": {
            "type": "string",
            "enum": ["hash", "semantic"]
          }
        }
      },
      "CodeItem": {
        "type": "object",
        "properties": {
          "nodeId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          },
          "span": {
            "type": "string"
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["nodeId", "type", "name", "terms"]
      },
      "CodeFileView": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "nodeId": {
            "type": "string"
          },
          "blobRef": {
            "type": "string"
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CodeItem"
            }
          }
        },
        "required": ["path", "nodeId", "terms", "items"]
      },
      "CodeItemView": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CodeItem"
          },
          {
            "type": "object",
            "properties": {
              "filePath": {
                "type": "string"
              },
              "callersIn": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CodeItem"
                }
              },
              "callsOut": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CodeItem"
                }
              }
            },
            "required": ["callersIn", "callsOut"]
          }
        ]
      },
      "RegionRule": {
        "type": "object",
        "properties": {
          "rule": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "reviewers": {},
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["rule", "paths"]
      },
      "CodeNeighborhood": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "terms": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": ["id", "label", "type", "terms"]
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "from": {
                  "type": "string"
                },
                "to": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              },
              "required": ["id", "from", "to", "type"]
            }
          }
        },
        "required": ["nodes", "edges"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/graphs": {
      "get": {
        "summary": "List registered graphs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All registered graphs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "graphs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GraphInfo"
                      }
                    }
                  },
                  "required": ["graphs"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Register a repo graph",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterGraphBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered graph entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphInfo"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid repo path"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/graphs/{id}": {
      "patch": {
        "summary": "Update graph settings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGraphBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated graph entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Graph not found"
          }
        }
      },
      "delete": {
        "summary": "Remove a registered graph",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Graph removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Graph not found"
          },
          "409": {
            "description": "Cannot remove the default graph"
          }
        }
      }
    },
    "/api/v1/code/tree": {
      "get": {
        "summary": "Code file tree",
        "description": "Nested directory tree of all indexed SourceFile nodes. Only available for repo graphs.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Nested file tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tree": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": ["tree"]
                }
              }
            }
          },
          "400": {
            "description": "Not a repo graph"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/code/file": {
      "get": {
        "summary": "Code file view",
        "description": "View a single SourceFile with its declared items. 404 if path is not indexed.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Repo-relative file path"
            },
            "required": true,
            "name": "path",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "File view with declared items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeFileView"
                }
              }
            }
          },
          "400": {
            "description": "Not a repo graph, or missing path param"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Path not indexed"
          }
        }
      }
    },
    "/api/v1/code/item/{nodeId}": {
      "get": {
        "summary": "Code item view",
        "description": "Single code item (function, class, etc.) with callers and callees.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "nodeId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Item view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeItemView"
                }
              }
            }
          },
          "400": {
            "description": "Not a repo graph"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Node not found"
          }
        }
      }
    },
    "/api/v1/code/regions": {
      "get": {
        "summary": "Policy region membership",
        "description": "Maps policy rules to the file paths they match via git_checklist.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Region rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rules": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RegionRule"
                      }
                    }
                  },
                  "required": ["rules"]
                }
              }
            }
          },
          "400": {
            "description": "Not a repo graph"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/code/neighborhood": {
      "get": {
        "summary": "Code neighborhood graph",
        "description": "Nodes and edges one hop from the given file path.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Repo-relative file path"
            },
            "required": true,
            "name": "path",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Neighborhood graph",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeNeighborhood"
                }
              }
            }
          },
          "400": {
            "description": "Not a repo graph or missing path param"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}